requirements.txt
pandas==2.2.0
playwright==1.41.1
python-dotenv==1.0.1
